Refactor unit tests to use new RecordingPreparedStatement location; remove old implementation
